chore(back-e2e): split maps.e2e-spec.ts
I'M SORRY FOR DUMPING THIS IN THE MIDDLE OF THIS PR BUT THIS FILE IS NUKING MY EDITOR.
So, Jest sucks. You're forced to run tests on a per-file basis, and we have to set up a Nest instance for at once.
If we ran Nest in the global test env, we can't access the instance in code, which we need to override providers in some cases.
So each separate file slows down CI so I try to avoid too many. Since this file is so monstrous and slowing down WebStorm,
I'm just splitting it in two. There's *probably* a better way of doing it, but honestly, Jest makes it difficult.
I'd rather not mess with it RN, so just doing this.
